% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_Matrix.R
\name{invertMatrix}
\alias{invertMatrix}
\title{Invert a Matrix}
\usage{
invertMatrix(M)
}
\arguments{
\item{M}{A square numeric matrix.}
}
\value{
A numeric matrix representing the inverse of the input matrix.
}
\description{
This function calculates the inverse of a given square matrix using C++ code.
}
\examples{
mat <- matrix(c(1, 2, 3, 4), nrow = 2)
invertMatrix(mat)
}
