% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Linear Model}
\usage{
linear_model(X, y, intercept = TRUE)
}
\arguments{
\item{X}{is a matrix of input features, each row represents an observation}

\item{y}{is nums of parameters you obtained from linear regression}

\item{intercept}{is a bool value denotes if model contain an intercept}
}
\value{
the result list of linear regression. Notice that, result model in the return list has removed the observation with NA.
}
\description{
This function is designed to obtain the result of linear regression
}
\examples{
#Use horsepower&weight to predict miles per gallon
X <- mtcars[, c("hp", "wt")]  # horsepower and weight
y <- mtcars$mpg               # miles per gallon

#Generate model using designed function
model <- linear_model(X,y,intercept=TRUE)
print(model)
}
