% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{Linear_regression}
\alias{Linear_regression}
\title{Linear Regression}
\usage{
Linear_regression(X, y, k = 0, intercept = TRUE, scale = FALSE)
}
\arguments{
\item{y}{is a matrix of output labels, each row represents an observation}

\item{x}{is a matrix of input features, each row represents an observation}
}
\value{
the matrix of parameters \eqn{\beta} forlinear regression
}
\description{
This function is designed to obtain the coefficient \eqn{\beta} from ridge regression
\deqn{Y = X\beta + e}
and minimize \eqn{||Y-X\beta||^2},
where \eqn{\beta} served as a predetermined hyperparamter to shift the diagonals of moment matrix.
More information about this method could be found in references.
}
\examples{
square(3)

}
\references{
Wikipedia contributors. (n.d.). Ridge regression. In Wikipedia. Retrieved [November, 11th, 2023], from https://en.wikipedia.org/wiki/Ridge_regression
}
