% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_t_test.R
\name{parameter_t_test}
\alias{parameter_t_test}
\title{Studnet-t test for the model generated by linear_model}
\usage{
parameter_t_test(model)
}
\arguments{
\item{model}{is the result of function linear_model()}
}
\value{
the list of Estimated value, standard error, t value, p-value of t-test for each coefficients
}
\description{
This function is designed to apply t-test to the result of fitted linear model coefficients
The corresponding t-test statistic is \eqn{t=\beta_j/SE(\beta_j)}
}
\examples{
# Load example dataset
data(mtcars)

#Use horsepower&weight to predict miles per gallon
X <- mtcars[, c("hp", "wt")]  # horsepower and weight
y <- mtcars$mpg               # miles per gallon

# Run the linear regression
model <- linear_model(X, y)
result <- parameter_t_test(model)

print(result)

}
\references{
Wikipedia contributors. (n.d.). Student's t-test. In Wikipedia, The Free Encyclopedia. Retrieved [11/13/2023], from https://en.wikipedia.org/wiki/Student%27s_t-test
}
