% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_prediction.R
\name{linear_prediction}
\alias{linear_prediction}
\title{Linear prediction}
\usage{
linear_prediction(X, b, intercept = TRUE)
}
\arguments{
\item{X}{is a matrix of input features, each row represents an observation}

\item{b}{is nums of parameters you obtained from linear regression}

\item{intercept}{is a bool value denotes if model contain an intercept}
}
\value{
the prediction \eqn{X\beta} for linear regression
}
\description{
This function is designed to obtain the prediction of linear regression
}
