% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{Linear_regression}
\alias{Linear_regression}
\title{Linear Regression}
\usage{
Linear_regression(X, y, intercept = TRUE)
}
\arguments{
\item{X}{is a matrix of input features, each row represents an observation}

\item{y}{is a matrix of output labels, each row represents an observation}

\item{intercept}{is a bool value indicates whether there is intercept in the model, default value is true}

\item{scale}{is a bool value indicates whether scale the input X in the model, default value is false}
}
\value{
the matrix of parameters \eqn{\beta} for linear regression
}
\description{
This function is designed to obtain the coefficient \eqn{\beta} from ridge regression
\deqn{Y = X\beta + e}
and minimize \eqn{||Y-X\beta||^2},
where \eqn{\beta} served as a desired parameter of model.
More information about this method could be found in references.
}
\examples{
# Load example dataset
data(mtcars)

# Prepare input data
X <- mtcars[, c("hp", "wt")]  # horsepower and weight
y <- mtcars$mpg               # miles per gallon

# Run the linear regression
result <- Linear_regression(X, y)

# Print the result
print(result)

}
\references{
Wikipedia contributors. (n.d.). Ridge regression. In Wikipedia. Retrieved [November, 11th, 2023], from https://en.wikipedia.org/wiki/Ridge_regression
}
